MIFOSADMIN-283, fixed Apache SOLR configuration.
This differs from commit 05e20f, whose comment sounds similar.
That commit contained changes being tested and deployed to
mifos_testing, but are not yet ready to deploy to mifos production. This
commit adds the Apache SOLR config change to the most recent commit that
was deployed to production, 4f8a1.

In the future, changes to be deployed to the test site should be pushed
to the repository on a branch other than master.

modified:   sites/all/modules/apachesolr/solrconfig.xml